refactor(model): renamed invited admin model
